requirements.txt
flask
flask-wtf
python-dotenv
openai
httpx